fix：修改专题组的kpi字段类型；feat：实验任务的count统计

+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/entity/subjectgroupmember/SubjectGroupMember.java
@@ -91,7 +91,7 @@ public class SubjectGroupMember extends BaseEntity {
      * 回款金额kpi
      */
     @Column(name = "kpi_return_fund", nullable = false )
-    private Long kpiReturnFund;
+    private BigDecimal kpiReturnFund;
 
     @OneToOne(fetch = FetchType.EAGER)
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
